+++ onessaye/onessaye/Models/POCO/SolidIngredient.cs
@@ -14,9 +14,9 @@ namespace onessaye.Models.POCO
         //Builder
         public SolidIngredient(string name, double amount, double unitPrice):base(name, amount, unitPrice) { }
         //Methods
-        public override double CalculCostIngredient()
+        public override float CalculCostIngredient()
         {
-            return UnitPrice * (Amount / 200);
+            return (float)(UnitPrice * (Amount / 200));
         }
         public override string ToString()
         {
